Implemented support for @keyword decorator
get_keyword_names populates the mapping from an 'embedded
argument name' to a 'function name'. Assumed is that
get_keyword_names is always called before run_keyword.

Tags passed to the decorator are added at the end of the keyword
documentation using Tags: tag1, tag2 syntax.

The test does not check if tag1 and tag2 have been added to the
keyword yet.

issue: #35