First step to allow running server on background. #22
Server is now started on a thread, but there is still now way to avoid
stopping it when the server is initialized.

All tests pass with this implementation, but test execution time
increases nearly 100%. That seems to be solely due to it taking more
time (0.6s vs. 0.1s on my setup) to wait for the server process to end
after using "Stop Remote Server" keyword. That shouldn't matter at all
in normal usage.