Remove broken async support.
Added ThreadingMixIn at some point when testing different approaches
to run server on background. It would add async support which would be
awesome, but unfortunately it doesn't work well with the current
system so intercept stdout/stderr. See issue #20 that requests this
async support for more information about the problems. The issue also
contains a potential solution, but it is too big task for 1.1 release.